organize: move showcase components into folder
