+++ frontend/src/components/showcase/ShowcaseSection.tsx
@@ -6,7 +6,7 @@ import {
 import { useEffect, useRef, useState } from "react";
 import { toast } from "react-toastify";
 import { PieceDTO } from "../dtos/dtos";
-import PieceAttributeCard from "./showcase/PieceAttributeCard";
+import PieceAttributeCard from "./PieceAttributeCard";
 
 // TODO: this is giving a weird pace to the size of some seconds
 function formatTime(time: number) {
